refactor: move check_keys.ts to scripts
